extract useCollections hook
